+++ src/helpers/bem-helper.js
@@ -1,5 +1,5 @@
 import * as ReactBemHelper from 'react-bem-helper'
 
 export const BemHelper = ReactBemHelper.withDefaults({
-  prefix: 'oy-',
+  prefix: 'ls-',
 })