+++ CHANGES.md
@@ -1,9 +1,15 @@
-# Changelog for Oyez Editor
+# Changelog for Lovely Editor
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and
 this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## 2018/07/05 [Unreleased][1]
+
+### Changed
+
+- *BREAKING*: changed all names starting like "oyez-" or "oy-" to either "lovely-" or "ls-"
 
 ## 2018/07/04 [Unreleased][1]